+++ example/src/module.ts
@@ -1,20 +1,39 @@
-import { defineNuxtModule } from '@nuxt/kit'
+import { defineNuxtModule, addPlugin, createResolver } from '@nuxt/kit'
 
+// Module options TypeScript interface definition
 export interface ModuleOptions {
   apiKey: string
 }
 
 export interface ModuleHooks {
-  'my-module:init': any
+  'my-module:init': () => void
+}
+
+export interface ModuleRuntimeHooks {
+  'my-module:runtime-hook': () => void
+}
+
+export interface ModulePublicRuntimeConfig {
+  NAME: string
+}
+
+export interface ModuleRuntimeConfig {
+  PRIVATE_NAME: string
 }
 
 export default defineNuxtModule<ModuleOptions>({
   meta: {
     name: 'my-module',
-    configKey: 'myModule'
+    configKey: 'myModule',
   },
+  // Default configuration options of the Nuxt module
   defaults: {
-    apiKey: null
+    apiKey: '',
   },
-  setup (_options, _nuxt) {}
+  setup(_options, _nuxt) {
+    const resolver = createResolver(import.meta.url)
+
+    // Do not add the extension since the `.ts` will be transpiled to `.mjs` after `npm run prepack`
+    addPlugin(resolver.resolve('./runtime/plugins/plugin'))
+  },
 })