+++ example/src/runtime/plugin.ts
@@ -1,7 +1,0 @@
-// @ts-ignore
-import { defineNuxtPlugin } from '#app'
-
-export default defineNuxtPlugin((_nuxtApp) => {
-  // eslint-disable-next-line no-console
-  console.log('Plugin by my-module!')
-})