+++ src/cli.ts
@@ -1,20 +1,29 @@
 #!/usr/bin/env node
+import { defineCommand, runMain } from 'citty'
+import type { CommandDef } from 'citty'
+import { consola } from 'consola'
+import { name, description, version } from '../package.json'
 
-/* eslint-disable no-console */
-import mri from 'mri'
-import { resolve } from 'pathe'
-import { buildModule } from './build'
+const _rDefault = (r: unknown) => (r && typeof r === 'object' && 'default' in r ? r.default : r) as Promise<CommandDef>
 
-async function main () {
-  const args = mri(process.argv.slice(2))
-  const rootDir = resolve(args._[0] || '.')
-  await buildModule({
-    rootDir,
-    stub: args.stub
-  })
-}
-
-main().catch((err) => {
-  console.error(err)
-  process.exit(1)
+const main = defineCommand({
+  meta: {
+    name,
+    description,
+    version,
+  },
+  subCommands: {
+    prepare: () => import('./commands/prepare').then(_rDefault),
+    build: () => import('./commands/build').then(_rDefault),
+  },
+  setup(context) {
+    // TODO: support 'default command' in citty?
+    const firstArg = context.rawArgs[0]
+    if (context.cmd.subCommands && !(firstArg in context.cmd.subCommands)) {
+      consola.warn('Please specify the `build` command explicitly. In a future version of `@nuxt/module-builder`, the implicit default build command will be removed.')
+      context.rawArgs.unshift('build')
+    }
+  },
 })
+
+runMain(main)