+++ src/index.ts
@@ -1,1 +1,2 @@
-export * from './build'
+export { default as build } from './commands/build'
+export { default as prepare } from './commands/prepare'