+++ test/__snapshots__/TestMe.vue
@@ -1,0 +1,17 @@
+<script>
+import { defineComponent } from "vue";
+export default defineComponent({
+  props: {
+    count: {
+      type: Number,
+      required: true
+    }
+  }
+});
+</script>
+
+<template>
+  <div>
+    {{ count }}
+  </div>
+</template>