+++ test/__snapshots__/plugin.d.ts
@@ -1,0 +1,6 @@
+declare const _default: import("#app").Plugin<{
+    injection: "injected";
+}> & import("#app").ObjectPlugin<{
+    injection: "injected";
+}>;
+export default _default;